requirements.txt
spotipy
python-dotenv
sv_ttk
tk
